+++ docs/CHANGELOG.md
@@ -5,9 +5,14 @@ For git tags, official releases are marked like "v1.0.0", while unofficial ones 
 
 ## [Unreleased]
 
+### Added
+
+* ``langFromDict`` and ``targetLangFromDict`` are now officially coming to the kernel utility functions.
+
 ### Fixed
 
 * When editing an existing text, surrounding spaces are no longer inserted ([#92](https://github.com/HugoFara/lwt/issues/92)).
+* Language code is better detected from translator url when editing an existing language.
 
 ## 2.8.1-fork (April 14 2023)